requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
backports.zoneinfo==0.2.1;python_version<"3.9"
breadability==0.1.20
build==1.2.2.post1
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
dataclasses-json==0.6.7
Django==4.2.28
djangorestframework==3.15.2
docopt==0.6.2
docopt-ng==0.9.0
exceptiongroup==1.3.1
faiss-cpu==1.8.0.post1
filelock==3.16.1
frozenlist==1.5.0
fsspec==2025.3.0
greenlet==3.1.1
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.4
langsmith==0.1.147
lxml==6.0.2
lxml_html_clean==0.4.3
MarkupSafe==2.1.5
marshmallow==3.22.0
mpmath==1.3.0
multidict==6.1.0
mypy_extensions==1.1.0
networkx==3.1
nltk==3.9.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
orjson==3.10.15
packaging==24.2
pillow==10.4.0
pip-tools==7.5.2
propcache==0.2.0
psycopg2-binary==2.9.10
pycountry==24.6.1
pydantic==2.10.6
pydantic_core==2.27.2
pyproject_hooks==1.2.0
python-dotenv==1.0.1
PyYAML==6.0.3
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
sniffio==1.3.1
SQLAlchemy==2.0.47
sqlparse==0.5.5
sumy==0.12.0
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.20.3
tomli==2.4.0
torch==2.4.1
tqdm==4.67.3
transformers==4.46.3
triton==3.0.0
typing-inspect==0.9.0
typing_extensions==4.13.2
urllib3==2.2.3
yarl==1.15.2
zipp==3.20.2
